# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: SwimManager Pro API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /teams:
    get:
      operationId: listTeams
      tags: [teams]
      summary: List all teams
      responses:
        "200":
          description: List of teams
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Team"
    post:
      operationId: createTeam
      tags: [teams]
      summary: Create a team
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TeamInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Team"
  /teams/{id}:
    get:
      operationId: getTeam
      tags: [teams]
      summary: Get team by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Team
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Team"
    patch:
      operationId: updateTeam
      tags: [teams]
      summary: Update team
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TeamUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Team"
    delete:
      operationId: deleteTeam
      tags: [teams]
      summary: Delete team
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── MEETS ─────────────────────────────────────────────────────────────────
...
  /timestandards:
    get:
      operationId: listTimeStandards
      tags: [timestandards]
      summary: List time standard sets
      responses:
        "200":
          description: Time standards
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TimeStandardSet"
    post:
      operationId: createTimeStandardSet
      tags: [timestandards]
      summary: Create a time standard set
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TimeStandardSetInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TimeStandardSet"
  /timestandards/{id}:
    delete:
      operationId: deleteTimeStandardSet
      tags: [timestandards]
      summary: Delete time standard set
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── WORKOUTS ──────────────────────────────────────────────────────────────
  /workouts:
    get:
      operationId: listWorkouts
      tags: [workouts]
      summary: List all workouts
      parameters:
        - name: teamId
          in: query
          schema:
            type: ["integer", "null"]
        - name: search
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Workouts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Workout"
    post:
      operationId: createWorkout
      tags: [workouts]
      summary: Create a workout
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkoutInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workout"
  /workouts/{id}:
    get:
      operationId: getWorkout
      tags: [workouts]
      summary: Get workout by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Workout
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workout"
    patch:
      operationId: updateWorkout
      tags: [workouts]
      summary: Update workout
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkoutUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workout"
    delete:
      operationId: deleteWorkout
      tags: [workouts]
      summary: Delete workout
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── BILLING ───────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Team:
      type: object
      required: [id, name]
      properties:
        id:
          type: integer
        name:
          type: string
        abbreviation:
          type: ["string", "null"]
        shortName:
          type: ["string", "null"]
        lsc:
          type: ["string", "null"]
        coachName:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        state:
          type: ["string", "null"]
        country:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        athleteCount:
          type: ["integer", "null"]
    TeamInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        abbreviation:
          type: string
        shortName:
          type: string
        lsc:
          type: string
        coachName:
          type: string
        address:
          type: string
        city:
          type: string
        state:
          type: string
        country:
          type: string
        phone:
          type: string
        email:
          type: string
    TeamUpdate:
      type: object
      properties:
        name:
          type: string
        abbreviation:
          type: string
        shortName:
          type: string
        lsc:
          type: string
        coachName:
          type: string
        address:
          type: string
        city:
          type: string
        state:
          type: string
        country:
          type: string
        phone:
          type: string
        email:
          type: string

    # ── MEET ──────────────────────────────────────────────────────────────
...
    TimeStandardSet:
      type: object
      required: [id, name]
      properties:
        id:
          type: integer
        name:
          type: string
        course:
          type: ["string", "null"]
        gender:
          type: ["string", "null"]
        ageGroup:
          type: ["string", "null"]
        standards:
          type: array
          items:
            $ref: "#/components/schemas/TimeStandard"
    TimeStandard:
      type: object
      required: [id, setId, label, distance, stroke, time]
      properties:
        id:
          type: integer
        setId:
          type: integer
        label:
          type: string
        distance:
          type: integer
        stroke:
          type: string
        gender:
          type: ["string", "null"]
        ageGroup:
          type: ["string", "null"]
        time:
          type: number
    TimeStandardSetInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        course:
          type: string
        gender:
          type: string
        ageGroup:
          type: string

    # ── WORKOUT ───────────────────────────────────────────────────────────
    Workout:
      type: object
      required: [id, title]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        teamId:
          type: ["integer", "null"]
        teamName:
          type: ["string", "null"]
        date:
          type: ["string", "null"]
        startTime:
          type: ["string", "null"]
        endTime:
          type: ["string", "null"]
        totalDistance:
          type: ["integer", "null"]
        course:
          type: ["string", "null"]
        isSample:
          type: boolean
        sets:
          type: array
          items:
            $ref: "#/components/schemas/WorkoutSet"
        createdAt:
          type: string
    WorkoutSet:
      type: object
      required: [id, workoutId, setOrder, description]
      properties:
        id:
          type: integer
        workoutId:
          type: integer
        setOrder:
          type: integer
        repetitions:
          type: ["integer", "null"]
        distance:
          type: ["integer", "null"]
        stroke:
          type: ["string", "null"]
        restInterval:
          type: ["string", "null"]
        description:
          type: string
        intensity:
          type: ["string", "null"]
    WorkoutInput:
      type: object
      required: [title]
      properties:
        title:
          type: string
        description:
          type: string
        teamId:
          type: integer
        date:
          type: string
        startTime:
          type: string
        endTime:
          type: string
        totalDistance:
          type: integer
        course:
          type: string
        isSample:
          type: boolean
        sets:
          type: array
          items:
            $ref: "#/components/schemas/WorkoutSetInput"
    WorkoutSetInput:
      type: object
      required: [setOrder, description]
      properties:
        setOrder:
          type: integer
        repetitions:
          type: integer
        distance:
          type: integer
        stroke:
          type: string
        restInterval:
          type: string
        description:
          type: string
        intensity:
          type: string
    WorkoutUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        teamId:
          type: integer
        date:
          type: string
        startTime:
          type: string
        endTime:
          type: string
        totalDistance:
          type: integer
        course:
          type: string
        isSample:
          type: boolean

    # ── BILLING ───────────────────────────────────────────────────────────
...