# lib/api-zod/src/generated/types/teamUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SwimManager Pro API
 * OpenAPI spec version: 0.1.0
 */

export interface TeamUpdate {
  name?: string;
  abbreviation?: string;
  shortName?: string;
  lsc?: string;
  coachName?: string;
  address?: string;
  city?: string;
  state?: string;
  country?: string;
  phone?: string;
  email?: string;
}