# lib/api-zod/src/generated/types/timeStandard.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SwimManager Pro API
 * OpenAPI spec version: 0.1.0
 */

export interface TimeStandard {
  id: number;
  setId: number;
  label: string;
  distance: number;
  stroke: string;
  /** @nullable */
  gender?: string | null;
  /** @nullable */
  ageGroup?: string | null;
  time: number;
}