# lib/api-zod/src/generated/types/workout.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SwimManager Pro API
 * OpenAPI spec version: 0.1.0
 */
import type { WorkoutSet } from './workoutSet';

export interface Workout {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  teamId?: number | null;
  /** @nullable */
  teamName?: string | null;
  /** @nullable */
  date?: string | null;
  /** @nullable */
  startTime?: string | null;
  /** @nullable */
  endTime?: string | null;
  /** @nullable */
  totalDistance?: number | null;
  /** @nullable */
  course?: string | null;
  isSample?: boolean;
  sets?: WorkoutSet[];
  createdAt?: string;
}